requirements.txt
dash==2.7.0
Flask==2.2.2
flask_sqlalchemy==3.0.2
mysql_connector_repackaged==0.3.1
pandas==1.4.3
plotly==5.9.0
PyMySQL==1.0.2
pyspark==3.1.2
SQLAlchemy==1.4.39
gunicorn==20.1.0
matplotlib==3.6.2
numpy==1.23.4
PyMySQL==1.0.2
scikit_learn==1.2.0
utm==0.7.0
